Convert JSON format to be easily readable
So far the JSON format contained strings of hexdata for the raw
variable data as well as the GUID. We are changing this to the
default GUID format for ease of use. For the same reason we are
changing the main data format from hex string to base64 encoded
as it is a format that is easy to generate inline from a the
comandline.

```
    {
        "name": "SecureBootEnable",
        "data": "AQ==",
        "guid": "f0a30bc7-af08-4556-99c4-001009c93a44",
        "attr": 3
    },
```
is easier to read and understand than

```
    {
        "name": "PK",
        "data": "a1 59 c0 a5 e4 94 a7 4a 87 b5 ab 15 5c 2b f0 72 2c 00 00 00 00 00 00 00 10 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00",
        "guid": "61 df e4 8b ca 93 d2 11 aa 0d 00 e0 98 03 2b 8c",
        "attr": 39
    },
```

Signed-off-by: Johanna Amélie Schander <mimoja@amazon.de>
Reviewed-by: Hendrik Borghorst <hborghor@amazon.de>