edk2: Calculate variable store size from file size
We have a (hard coded for now) value for the overall edk2 store size.
However, that store is only a container for the actual variable store
which has its own size field which we did not derive from the overall
size.

That meant that you could get the incorrect impression that you can
choose a different file size while in reality, the resulting store would
not work.

Fix that by calculating the embedded store file size value depending on
the overall store size.

Reported-by: Benjamin Herrenschmidt <benh@amazon.com>
Signed-off-by: Alexander Graf <graf@amazon.com>