Allow custom sized EDK2 store via new filesize subparam to the output format
We have a static and compiletime set value for the overall store size in an
specific EDK2 build.

We are adding the capability to pass the final variable-store filesize in KiB
to the EDK2 output formatter to allow for all kinds of EDK2 configurations.
Expected values are 256, 512 and 1024 but we are not limiting the output format
to those.

Reported-by: Benjamin Herrenschmidt <benh@amazon.com>

CC: Alex Graf <graf@amazon.com>
Cc-Team: kaos-brimstone <kaos-brimstone@amazon.com>
Signed-off-by: Johanna Amélie Schander <mimoja@amazon.de>