Cleanup (json) output format and introduce a version field
Up to this point our json format went through some changes.
As we are approaching ABI stability we are now introducing
a version field to ensure that going forward we can update the
format without breaking existing json files. The format
that we used up to this point is implicitly considered V1 while
this commit introduces the V2.

CC: Alex Graf <graf@amazon.com>
Cc-Team: kaos-brimstone <kaos-brimstone@amazon.com>

Signed-off-by: Johanna Amélie Schander <mimoja@amazon.de>