json: Convert JSONVar to UEFIVar when reading the varstore
The JSONVar dictionary variables are of a different format if
they are read from json vs if they are converted from the generic
UEFIVar. To ensure we have always a known format we enforce an UEFIVar
format after reading. This requires us to also check the existence of
all entries and convert them to a proper format if we are writing json.

Signed-off-by: Johanna Amélie Schander <mimoja@amazon.de>