edk2.py: Use length from firmware volume header (Fixes #9)
The firmware volume length might not cover the full input length, thus
resulting in an unexpected (and likely incorrect) value being written to
the output firmware volume header.

Resolve this by using the length from the input firmware volume header
instead, verifying that this length is plausible for the input file.

Signed-off-by: Mate Kukri <mate.kukri@canonical.com>