efivarfs: Skip authenticated variables
With efivarfs, we only get to see the output of GET_VARIABLE which does not
return the contents of the validation certificate for authentication. So
when we read an authenticated variable, we would read it without any
certificate that anyone could ever use to change it again.

Instead of leaving behind immutable authenticated variables, let's skip all
of them so they don't slip in by accident and warn the user.

Signed-off-by: Alexander Graf <graf@amazon.com>