Add input support for efivarfs
We want to create a UEFI variable store from the variables we can see at
runtime in a running OS, so we can recreate the variable store of a live
system. To do that, this patch adds support to use efivarfs - the file
system Linux exposes to access the runtime UEFI variable store - as input.

Originally-authored-by: Michael Yuniverg <michael.yuniverg@cloudendure.com>
[graf: Minor refactoring to avoid regexp and reduce side effects]
Signed-off-by: Alexander Graf <graf@amazon.com>